requirements.txt
allure-pytest==2.14.0
allure-python-commons==2.14.0
attrs==25.3.0
colorama==0.4.6
iniconfig==2.1.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.5
